+++ application/src/test/java/com/ericsson/bss/cassandra/ecchronos/application/config/TestDefaultConfig.java
@@ -14,8 +14,16 @@
  */
 package com.ericsson.bss.cassandra.ecchronos.application.config;
 
-import com.fasterxml.jackson.core.exc.StreamReadException;
-import com.fasterxml.jackson.databind.DatabindException;
+import com.ericsson.bss.cassandra.ecchronos.application.config.connection.ConnectionConfig;
+import com.ericsson.bss.cassandra.ecchronos.application.config.connection.DistributedJmxConnection;
+import com.ericsson.bss.cassandra.ecchronos.application.config.connection.RetryPolicyConfig;
+import com.ericsson.bss.cassandra.ecchronos.application.config.repair.FileBasedRepairConfiguration;
+import com.ericsson.bss.cassandra.ecchronos.application.config.repair.GlobalRepairConfig;
+import com.ericsson.bss.cassandra.ecchronos.application.config.repair.Interval;
+import com.ericsson.bss.cassandra.ecchronos.application.config.repair.Priority;
+import com.ericsson.bss.cassandra.ecchronos.application.providers.AgentJmxConnectionProvider;
+import com.ericsson.bss.cassandra.ecchronos.utils.enums.repair.RepairHistoryProvider;
+import com.ericsson.bss.cassandra.ecchronos.utils.enums.repair.RepairType;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
 import org.junit.Before;
@@ -24,17 +32,19 @@ import org.junit.Test;
 import java.io.File;
 import java.io.IOException;
 import java.util.concurrent.TimeUnit;
-
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
 
 public class TestDefaultConfig
 {
     private static final String DEFAULT_AGENT_FILE_NAME = "nothing_set.yml";
+    private static final TimeUnit TIME_UNIT_IN_SECONDS = TimeUnit.SECONDS;
     private static Config config;
-
+    private static GlobalRepairConfig repairConfig;
 
     @Before
-    public void setup() throws StreamReadException, DatabindException, IOException
+    public void setup() throws IOException
     {
         ClassLoader classLoader = Thread.currentThread().getContextClassLoader();
 
@@ -43,6 +53,7 @@ public class TestDefaultConfig
         ObjectMapper objectMapper = new ObjectMapper(new YAMLFactory());
 
         config = objectMapper.readValue(file, Config.class);
+        repairConfig = config.getRepairConfig();
 
     }
 
@@ -54,4 +65,105 @@ public class TestDefaultConfig
         assertThat(connectionDelay.getTime()).isEqualTo(60l);
     }
 
+    @Test
+    public void testRetryPolicyConfigWhenNothingSet()
+    {
+        ConnectionConfig connection = config.getConnectionConfig();
+        DistributedJmxConnection distributedJmxConnection = connection.getJmxConnection();
+        Class<?> providerClass = distributedJmxConnection.getProviderClass();
+        assertThat(providerClass).isEqualTo(AgentJmxConnectionProvider.class);
+        RetryPolicyConfig retryPolicyConfig = distributedJmxConnection.getRetryPolicyConfig();
+        assertNotNull(retryPolicyConfig);
+        assertThat(5).isEqualTo(retryPolicyConfig.getMaxAttempts());
+        assertThat(5000).isEqualTo(retryPolicyConfig.getRetryDelay().getStartDelay());
+        assertThat(30000).isEqualTo(retryPolicyConfig.getRetryDelay().getMaxDelay());
+        assertThat(TIME_UNIT_IN_SECONDS).isEqualTo(retryPolicyConfig.getRetryDelay().getUnit());
+        assertThat(86400000).isEqualTo(retryPolicyConfig.getRetrySchedule().getInitialDelay());
+        assertThat(86400000).isEqualTo(retryPolicyConfig.getRetrySchedule().getFixedDelay());
+        assertThat(TIME_UNIT_IN_SECONDS).isEqualTo(retryPolicyConfig.getRetrySchedule().getUnit());
+    }
+
+    @Test
+    public void testRepairProvider()
+    {
+        Class<?> providerClass = repairConfig.getRepairConfigurationClass();
+        assertEquals(FileBasedRepairConfiguration.class, providerClass);
+    }
+
+    @Test
+    public void testRepairInterval()
+    {
+        Interval repairInterval = repairConfig.getRepairInterval();
+        assertThat(repairInterval.getUnit()).isEqualTo(TimeUnit.DAYS);
+        assertThat(repairInterval.getTime()).isEqualTo(7L);
+    }
+
+    @Test
+    public void testRepairInitialDelay()
+    {
+        Interval repairInitialDelay = repairConfig.getInitialDelay();
+        assertThat(repairInitialDelay.getUnit()).isEqualTo(TimeUnit.DAYS);
+        assertThat(repairInitialDelay.getTime()).isEqualTo(1L);
+    }
+
+    @Test
+    public void testRepairPriority()
+    {
+        Priority priority = repairConfig.getPriority();
+        assertEquals(priority.getPriorityGranularityUnit(), TimeUnit.HOURS);
+    }
+
+    @Test
+    public void testRepairUnwindRatio()
+    {
+        Double repairUnwindRatio = repairConfig.asRepairConfiguration().getRepairUnwindRatio();
+        assertEquals(repairUnwindRatio, Double.valueOf(0.0));
+    }
+
+    @Test
+    public void testRepairHistoryLookback()
+    {
+        Interval repairHistoryLookback = repairConfig.getRepairHistoryLookback();
+        assertThat(repairHistoryLookback.getUnit()).isEqualTo(TimeUnit.DAYS);
+        assertThat(repairHistoryLookback.getTime()).isEqualTo(30L);
+    }
+
+    @Test
+    public void testRepairSizeTarget()
+    {
+        long repairSizeTarget = repairConfig
+                .asRepairConfiguration().getTargetRepairSizeInBytes();
+        assertEquals(repairSizeTarget, Long.MAX_VALUE);
+    }
+
+    @Test
+    public void testRepairIgnoreTWCSTables()
+    {
+        boolean repairIgnoreTWCSTables = repairConfig.getIgnoreTWCSTables();
+        assertThat(repairIgnoreTWCSTables).isFalse();
+    }
+
+    @Test
+    public void testRepairHistoryProvider()
+    {
+        RepairHistoryProvider repairHistoryProvider = repairConfig.getRepairHistory().getProvider();
+        String keyspace = repairConfig.getRepairHistory().getKeyspaceName();
+        assertEquals(repairHistoryProvider, RepairHistoryProvider.ECC);
+        assertEquals(keyspace, "ecchronos");
+    }
+
+    @Test
+    public void testRepairBackoff()
+    {
+        Interval repairBackoff = repairConfig.getBackoff();
+        assertThat(repairBackoff.getUnit()).isEqualTo(TimeUnit.MINUTES);
+        assertThat(repairBackoff.getTime()).isEqualTo(30L);
+    }
+
+    @Test
+    public void testRepairType()
+    {
+        RepairType repairType = repairConfig.getRepairType();
+        assertThat(repairType).isEqualTo(RepairType.VNODE);
+    }
 }